Implement rollback to previous secrets version
"Upgrading" to the previously deployed version is special because the
secret already exists. It is kept around because stateful sets roll over
one instance at a time, and any failing older instances will be restarted
with the previous pod version until it is their time to upgrade.

This also means that pods will not wait for the secrets generator to finish.

Therefore we simply swap the current and previous secret names in the
configmap and update that.

Note that in general rollback is not supported: it is possible that the
current version was migrating services that are now no longer compatible
with the previous version of the software.  This change is no general
endorsement of rollbacks, but just tries to help out users who otherwise
know exactly what they are doing.