Add -isInstall option
On initial install the config map and secrets are deleted before they are
created again. This helps in case the user had installed and deleted the
software via helm before, but didn't manually remove the config map and
secrets. These are left behind because they are not accounted for in the
helm charts.