Allow template expansion for all variable options
Previously this was hardcoded for the `alternative_names` list
only, but now is available anywhere under the `options` key.

In addition this commit adds a --set commandline option to specify
additional template environment variables.

Unrelated, this commit also removes the `common_name` setting,
which is not used anywhere in the cert generation code.

[jsc#CAP-587]